+++ packages/evm/tasks/deploy/adapters/index.ts
@@ -6,6 +6,8 @@ import "./connext"
 import "./dendreth"
 import "./hyperlane"
 import "./layerzero"
+import "./optimism"
 import "./sygma"
 import "./telepathy"
+import "./wormhole"
 import "./zetachain"
