+++ packages/evm/tasks/deploy/index.ts
@@ -3,10 +3,7 @@ import { HardhatRuntimeEnvironment } from "hardhat/types"
 
 import "./adapters"
 import "./hashi"
-import "./optimism"
 import "./replay"
-import "./sygma"
-import "./wormhole"
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const verify = async (hre: HardhatRuntimeEnvironment, contract: Contract, constructorArguments: any = []) => {
